Review comments: Refactored code to just have Size based splitter
