+++ engine/src/main/java/com/google/android/fhir/sync/Sync.kt
@@ -27,7 +27,7 @@ import com.google.android.fhir.FhirEngine
 import com.google.android.fhir.FhirEngineProvider
 import com.google.android.fhir.sync.download.DownloaderImpl
 import com.google.android.fhir.sync.upload.BundleUploader
-import com.google.android.fhir.sync.upload.Splitter
+import com.google.android.fhir.sync.upload.LocalChangesPaginator
 import com.google.android.fhir.sync.upload.TransactionBundleGenerator
 import org.hl7.fhir.r4.model.ResourceType
 
@@ -56,7 +56,7 @@ object Sync {
           BundleUploader(
             it,
             TransactionBundleGenerator.getDefault(),
-            Splitter.create(uploadConfiguration)
+            LocalChangesPaginator.create(uploadConfiguration)
           ),
           DownloaderImpl(it, downloadManager),
           resolver
